+++ test/rxjs/rxjs.spec.ts
@@ -5,276 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-
-import * as Rx from 'rxjs/Rx';
-
-/**
- * The point of these tests, is to ensure that all callbacks execute in the Zone which was active
- * when the callback was passed into the Rx.
- *
- * The implications are:
- * - Observable callback passed into `Observable` executes in the same Zone as when the
- *   `new Observable` was invoked.
- * - The subscription callbacks passed into `subscribe` execute in the same Zone as when the
- *   `subscribe` method was invoked.
- * - The operator callbacks passe into `map`, etc..., execute in the same Zone as when the
- *   `operator` (`lift`) method was invoked.
- */
-describe('Zone interaction', () => {
-  it('should run methods in the zone of declaration', () => {
-    const log: string[] = [];
-    const constructorZone: Zone = Zone.current.fork({name: 'Constructor Zone'});
-    const subscriptionZone: Zone = Zone.current.fork({name: 'Subscription Zone'});
-    let subscriber: any = null;
-    const observable: any = constructorZone.run(() => new Rx.Observable((_subscriber: any) => {
-      subscriber = _subscriber;
-      log.push('setup');
-      expect(Zone.current.name).toEqual(constructorZone.name);
-      return () => {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        log.push('cleanup');
-      };
-    }));
-    subscriptionZone.run(
-        () => observable.subscribe(
-            () => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('next');
-            },
-            (): any => null,
-            () => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('complete');
-            }));
-    subscriber.next('MyValue');
-    subscriber.complete();
-
-    expect(log).toEqual(['setup', 'next', 'complete', 'cleanup']);
-    log.length = 0;
-
-    subscriptionZone.run(() => observable.subscribe((): any => null, () => {
-      expect(Zone.current.name).toEqual(subscriptionZone.name);
-      log.push('error');
-    }, (): any => null));
-    subscriber.next('MyValue');
-    subscriber.error('MyError');
-
-    expect(log).toEqual(['setup', 'error', 'cleanup']);
-  });
-
-  it('should run methods in the zone of declaration when nexting synchronously', () => {
-    const log: string[] = [];
-    const rootZone: Zone = Zone.current;
-    const constructorZone: Zone = Zone.current.fork({name: 'Constructor Zone'});
-    const subscriptionZone: Zone = Zone.current.fork({name: 'Subscription Zone'});
-    const observable: any = constructorZone.run(() => new Rx.Observable((subscriber: any) => {
-      // Execute the `next`/`complete` in different zone, and assert that
-      // correct zone
-      // is restored.
-      rootZone.run(() => {
-        subscriber.next('MyValue');
-        subscriber.complete();
-      });
-      return () => {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        log.push('cleanup');
-      };
-    }));
-
-    subscriptionZone.run(
-        () => observable.subscribe(
-            () => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('next');
-            },
-            (): any => null,
-            () => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('complete');
-            }));
-
-    expect(log).toEqual(['next', 'complete', 'cleanup']);
-  });
-
-  it('should run operators in the zone of declaration', () => {
-    const log: string[] = [];
-    const rootZone: Zone = Zone.current;
-    const constructorZone: Zone = Zone.current.fork({name: 'Constructor Zone'});
-    const operatorZone: Zone = Zone.current.fork({name: 'Operator Zone'});
-    const subscriptionZone: Zone = Zone.current.fork({name: 'Subscription Zone'});
-    let observable: any = constructorZone.run(() => new Rx.Observable((subscriber: any) => {
-      // Execute the `next`/`complete` in different zone, and assert that
-      // correct zone
-      // is restored.
-      rootZone.run(() => {
-        subscriber.next('MyValue');
-        subscriber.complete();
-      });
-      return () => {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        log.push('cleanup');
-      };
-    }));
-
-    observable = operatorZone.run(() => observable.map((value: any) => {
-      expect(Zone.current.name).toEqual(operatorZone.name);
-      log.push('map: ' + value);
-      return value;
-    }));
-
-    subscriptionZone.run(
-        () => observable.subscribe(
-            () => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('next');
-            },
-            (e: any) => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('error: ' + e);
-            },
-            () => {
-              expect(Zone.current.name).toEqual(subscriptionZone.name);
-              log.push('complete');
-            }));
-
-    expect(log).toEqual(['map: MyValue', 'next', 'complete', 'cleanup']);
-  });
-
-  it('should run subscribe in zone of declaration with Observable.create', () => {
-    const log: string[] = [];
-    const constructorZone: Zone = Zone.current.fork({name: 'Constructor Zone'});
-    let observable: any = constructorZone.run(() => Rx.Observable.create((subscriber: any) => {
-      subscriber.next(1);
-      subscriber.complete();
-      return () => {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        log.push('cleanup');
-      };
-    }));
-
-    observable.subscribe(() => {
-      log.push('next');
-    });
-
-    expect(log).toEqual(['next', 'cleanup']);
-  });
-
-  it('should run in the zone when subscribe is called to the same Subject', () => {
-    const log: string[] = [];
-    const constructorZone: Zone = Zone.current.fork({name: 'Constructor Zone'});
-    const subscriptionZone1: Zone = Zone.current.fork({name: 'Subscription Zone 1'});
-    const subscriptionZone2: Zone = Zone.current.fork({name: 'Subscription Zone 2'});
-
-    let subject: any;
-
-    constructorZone.run(() => {
-      subject = new Rx.Subject();
-    });
-
-    let subscription1: any;
-    let subscription2: any;
-
-    subscriptionZone1.run(() => {
-      subscription1 = subject.subscribe(
-          () => {
-            expect(Zone.current.name).toEqual(subscriptionZone1.name);
-            log.push('next1');
-          },
-          () => {},
-          () => {
-            expect(Zone.current.name).toEqual(subscriptionZone1.name);
-            log.push('complete1');
-          });
-    });
-
-    subscriptionZone2.run(() => {
-      subscription2 = subject.subscribe(
-          () => {
-            expect(Zone.current.name).toEqual(subscriptionZone2.name);
-            log.push('next2');
-          },
-          () => {},
-          () => {
-            expect(Zone.current.name).toEqual(subscriptionZone2.name);
-            log.push('complete2');
-          });
-    });
-
-    subject.next(1);
-    subject.complete();
-
-    expect(log).toEqual(['next1', 'next2', 'complete1', 'complete2']);
-  });
-
-  it('bindCallback func callback should run in the correct zone', (done: any) => {
-    let log: string[] = [];
-    const constructorZone: Zone = Zone.current.fork({name: 'Constructor Zone'});
-    const triggerZone: Zone = Zone.current.fork({name: 'Trigger Zone'});
-    const subscriptionZone: Zone = Zone.current.fork({name: 'Subscription Zone'});
-
-    let func: any;
-    let boundFunc: any;
-    let observable: any;
-
-    constructorZone.run(() => {
-      func = function(arg0: any, callback: Function) {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        callback(arg0);
-      };
-      boundFunc = Rx.Observable.bindCallback(func);
-      observable = boundFunc('test');
-    });
-
-    subscriptionZone.run(() => {
-      observable.subscribe((arg: any) => {
-        expect(Zone.current.name).toEqual(subscriptionZone.name);
-        log.push('next' + arg);
-      });
-    });
-
-    expect(log).toEqual(['nexttest']);
-    log.length = 0;
-
-    constructorZone.run(() => {
-      func = function(arg0: any, callback: Function) {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        callback(arg0);
-      };
-      boundFunc = Rx.Observable.bindCallback(func, (arg: any) => {
-        return 'selector' + arg;
-      });
-      observable = boundFunc('test');
-    });
-
-    subscriptionZone.run(() => {
-      observable.subscribe((arg: any) => {
-        expect(Zone.current.name).toEqual(subscriptionZone.name);
-        log.push('next' + arg);
-      });
-    });
-
-    expect(log).toEqual(['nextselectortest']);
-    log.length = 0;
-
-    constructorZone.run(() => {
-      func = function(arg0: any, callback: Function) {
-        expect(Zone.current.name).toEqual(constructorZone.name);
-        callback(arg0);
-      };
-      boundFunc = Rx.Observable.bindCallback(func, null, Rx.Scheduler.async);
-      observable = boundFunc('test');
-    });
-
-    subscriptionZone.run(() => {
-      observable.subscribe((arg: any) => {
-        expect(Zone.current.name).toEqual(subscriptionZone.name);
-        log.push('next' + arg);
-        done();
-      });
-    });
-
-    expect(log).toEqual([]);
-  });
-
-
-});
+import './rxjs.common.spec';
+import './rxjs.bindCallback.spec';
+import './rxjs.bindNodeCallback.spec';
+import './rxjs.combineLatest.spec';
