+++ test/test-util.ts
@@ -86,3 +86,13 @@ export function isSupportSetErrorStack() {
 }
 
 (isSupportSetErrorStack as any).message = 'supportSetErrorStack';
+
+export function asyncTest(testFn: Function, zone: Zone = Zone.current) {
+  const AsyncTestZoneSpec = (Zone as any)['AsyncTestZoneSpec'];
+  return (done: Function) => {
+    let asyncTestZone: Zone = zone.fork(new AsyncTestZoneSpec(done, (error: Error) => {
+      fail(error);
+    }, 'asyncTest'));
+    asyncTestZone.run(testFn, this, [done]);
+  };
+}
